refactor: extract shared tsconfig JSONC-stripping helper into zfb-plugin-resolver
Both crates/zfb/src/commands/build.rs and crates/zfb-islands/src/scanner.rs
contained independent tsconfig compilerOptions.paths / extends-chain walkers.
The scanner variant lacked JSONC comment stripping, causing a silent parse
failure on any tsconfig.json with // or /* */ comments — a latent bug for
islands alias resolution.

Extract a single shared helper (read_tsconfig_paths / read_tsconfig_paths_into_map)
into zfb-plugin-resolver with JSONC stripping and file-based extends resolution
(supporting any tsconfig filename, e.g. tsconfig.base.json). Convert both call
sites to use the shared helper. Move build.rs read_tsconfig_paths tests to
zfb-plugin-resolver and add islands-scanner JSONC regression test.

Fixes: #901
